refactor: refactored into a seperate pypi module
