+++ src/install_pypi.rs
@@ -37,7 +37,7 @@ use crate::{
     consts::{DEFAULT_PYPI_INDEX_URL, PIXI_UV_INSTALLER, PROJECT_MANIFEST},
     lock_file::UvResolutionContext,
     prefix::Prefix,
-    project::manifest::{pypi_options::PypiOptions, SystemRequirements},
+    project::manifest::{pypi::pypi_options::PypiOptions, SystemRequirements},
     pypi_tags::{get_pypi_tags, is_python_record},
     uv_reporter::{UvReporter, UvReporterOptions},
 };
