+++ src/lock_file/satisfiability.rs
@@ -1,7 +1,7 @@
 use super::{PypiRecord, PypiRecordsByName, RepoDataRecordsByName};
 use crate::project::grouped_environment::GroupedEnvironment;
 use crate::project::has_features::HasFeatures;
-use crate::project::manifest::python::AsPep508Error;
+use crate::project::manifest::pypi::pypi_requirement::AsPep508Error;
 use crate::{project::Environment, pypi_marker_env::determine_marker_environment};
 use itertools::Itertools;
 use miette::Diagnostic;
