+++ src/project/manifest/pypi/mod.rs
@@ -1,0 +1,5 @@
+pub mod pypi_options;
+pub mod pypi_requirement;
+pub mod pypi_requirement_types;
+
+pub use pypi_requirement_types::{GitRev, PyPiPackageName, VersionOrStar};
